Port changes (safety and auth)

+++ webapi/Models/Request/EditChatParameters.cs
@@ -27,5 +27,5 @@ public class EditChatParameters
     /// The higher this value, the more the system will rely on long term memory by lowering
     /// the relevance threshold of long term memory and increasing the threshold score of working memory.
     /// </summary>
-    public double? MemoryBalance { get; set; }
+    public float? MemoryBalance { get; set; }
 }
